small refactor in tests
