+++ core/src/test/scala/tech/beshu/ror/unit/acl/blocks/rules/indices/BaseIndicesRuleTests.scala
@@ -1,0 +1,255 @@
+/*
+ *    This file is part of ReadonlyREST.
+ *
+ *    ReadonlyREST is free software: you can redistribute it and/or modify
+ *    it under the terms of the GNU General Public License as published by
+ *    the Free Software Foundation, either version 3 of the License, or
+ *    (at your option) any later version.
+ *
+ *    ReadonlyREST is distributed in the hope that it will be useful,
+ *    but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *    GNU General Public License for more details.
+ *
+ *    You should have received a copy of the GNU General Public License
+ *    along with ReadonlyREST.  If not, see http://www.gnu.org/licenses/
+ */
+package tech.beshu.ror.unit.acl.blocks.rules.indices
+
+import eu.timepit.refined.auto._
+import cats.data.NonEmptySet
+import com.softwaremill.sttp.Method
+import eu.timepit.refined.api.Refined
+import eu.timepit.refined.numeric.Positive
+import eu.timepit.refined.types.string.NonEmptyString
+import org.scalamock.scalatest.MockFactory
+import org.scalatest.{Assertion, Succeeded}
+import org.scalatest.matchers.should.Matchers
+import org.scalatest.wordspec.AnyWordSpec
+import tech.beshu.ror.accesscontrol.blocks.BlockContext.{FilterableMultiRequestBlockContext, GeneralIndexRequestBlockContext, TemplateRequestBlockContext}
+import tech.beshu.ror.accesscontrol.blocks.BlockContext.MultiIndexRequestBlockContext.Indices
+import tech.beshu.ror.accesscontrol.blocks.metadata.UserMetadata
+import tech.beshu.ror.accesscontrol.blocks.rules.Rule.RuleResult.{Fulfilled, Rejected}
+import tech.beshu.ror.accesscontrol.blocks.rules.Rule.RuleResult.Rejected.Cause
+import tech.beshu.ror.accesscontrol.blocks.rules.elasticsearch.indices.IndicesRule
+import tech.beshu.ror.accesscontrol.blocks.variables.runtime.{RuntimeMultiResolvableVariable, RuntimeResolvableVariableCreator}
+import tech.beshu.ror.accesscontrol.blocks.variables.runtime.RuntimeMultiResolvableVariable.AlreadyResolved
+import tech.beshu.ror.accesscontrol.blocks.variables.runtime.RuntimeResolvableVariable.Convertible.AlwaysRightConvertible
+import tech.beshu.ror.accesscontrol.blocks.variables.transformation.{SupportedVariablesFunctions, TransformationCompiler}
+import tech.beshu.ror.accesscontrol.domain.ClusterIndexName.Remote.ClusterName
+import tech.beshu.ror.accesscontrol.domain.DataStreamName.FullRemoteDataStreamWithAliases
+import tech.beshu.ror.accesscontrol.domain.Template.{ComponentTemplate, IndexTemplate, LegacyTemplate}
+import tech.beshu.ror.accesscontrol.domain.{Action, ClusterIndexName, FullRemoteIndexWithAliases, IndexAttribute, IndexName, Template, TemplateName, TemplateOperation}
+import tech.beshu.ror.accesscontrol.matchers.RandomBasedUniqueIdentifierGenerator
+import tech.beshu.ror.mocks.{MockFilterableMultiRequestContext, MockGeneralIndexRequestContext, MockRequestContext, MockTemplateRequestContext}
+import tech.beshu.ror.utils.TestsUtils.{clusterIndexName, fullDataStreamName, fullIndexName, fullLocalIndexWithAliases, indexPattern}
+import tech.beshu.ror.utils.uniquelist.UniqueNonEmptyList
+import monix.execution.Scheduler.Implicits.global
+
+abstract class BaseIndicesRuleTests extends AnyWordSpec with Matchers with MockFactory {
+
+  protected def assertMatchRuleForIndexRequest(configured: NonEmptySet[RuntimeMultiResolvableVariable[ClusterIndexName]],
+                                             requestIndices: Set[ClusterIndexName],
+                                             modifyRequestContext: MockGeneralIndexRequestContext => MockGeneralIndexRequestContext = identity,
+                                             found: Set[ClusterIndexName]): Assertion =
+    assertRuleForIndexRequest(configured, requestIndices, isMatched = true, modifyRequestContext, found)
+
+  protected def assertNotMatchRuleForIndexRequest(configured: NonEmptySet[RuntimeMultiResolvableVariable[ClusterIndexName]],
+                                                requestIndices: Set[ClusterIndexName],
+                                                modifyRequestContext: MockGeneralIndexRequestContext => MockGeneralIndexRequestContext = identity): Assertion =
+    assertRuleForIndexRequest(configured, requestIndices, isMatched = false, modifyRequestContext, Set.empty)
+
+  private def assertRuleForIndexRequest(configuredValues: NonEmptySet[RuntimeMultiResolvableVariable[ClusterIndexName]],
+                                        requestIndices: Set[ClusterIndexName],
+                                        isMatched: Boolean,
+                                        modifyRequestContext: MockGeneralIndexRequestContext => MockGeneralIndexRequestContext,
+                                        found: Set[ClusterIndexName]) = {
+    val rule = createIndicesRule(configuredValues)
+    val requestContext = modifyRequestContext apply MockRequestContext.indices
+      .copy(
+        filteredIndices = requestIndices,
+        action = Action("indices:data/read/search"),
+        isReadOnlyRequest = true,
+        allIndicesAndAliases = Set(
+          fullLocalIndexWithAliases(fullIndexName("test1")),
+          fullLocalIndexWithAliases(fullIndexName("test2")),
+          fullLocalIndexWithAliases(fullIndexName("test3")),
+          fullLocalIndexWithAliases(fullIndexName("test4")),
+          fullLocalIndexWithAliases(fullIndexName("test5"))
+        )
+      )
+    val blockContext = GeneralIndexRequestBlockContext(
+      requestContext,
+      UserMetadata.from(requestContext),
+      Set.empty,
+      List.empty,
+      requestIndices,
+      Set.empty
+    )
+    rule.check(blockContext).runSyncStep shouldBe Right {
+      if (isMatched) {
+        Fulfilled(GeneralIndexRequestBlockContext(
+          requestContext = requestContext,
+          userMetadata = UserMetadata.from(requestContext),
+          responseHeaders = Set.empty,
+          responseTransformations = List.empty,
+          filteredIndices = found,
+          allAllowedIndices = configuredValues
+            .toNonEmptyList.toList
+            .collect { case a: AlreadyResolved[ClusterIndexName] => a }
+            .flatMap(_.value.toList)
+            .toSet
+        ))
+      } else {
+        Rejected(Some(Cause.IndexNotFound))
+      }
+    }
+  }
+
+  protected def assertMatchRuleForMultiIndexRequest(configured: NonEmptySet[RuntimeMultiResolvableVariable[ClusterIndexName]],
+                                                  indexPacks: List[Indices],
+                                                  modifyRequestContext: MockFilterableMultiRequestContext => MockFilterableMultiRequestContext = identity,
+                                                  allowed: List[Indices]): Assertion = {
+    assertRuleForMultiForIndexRequest(configured, indexPacks, isMatched = true, modifyRequestContext, allowed)
+  }
+
+  protected def assertNotMatchRuleForMultiIndexRequest(configured: NonEmptySet[RuntimeMultiResolvableVariable[ClusterIndexName]],
+                                                     indexPacks: List[Indices],
+                                                     modifyRequestContext: MockFilterableMultiRequestContext => MockFilterableMultiRequestContext = identity): Assertion = {
+    assertRuleForMultiForIndexRequest(configured, indexPacks, isMatched = false, modifyRequestContext, List.empty)
+  }
+
+  private def assertRuleForMultiForIndexRequest(configuredValues: NonEmptySet[RuntimeMultiResolvableVariable[ClusterIndexName]],
+                                                indexPacks: List[Indices],
+                                                isMatched: Boolean,
+                                                modifyRequestContext: MockFilterableMultiRequestContext => MockFilterableMultiRequestContext,
+                                                allowed: List[Indices]) = {
+    val rule = new IndicesRule(
+      settings = IndicesRule.Settings(configuredValues, mustInvolveIndices = false),
+      identifierGenerator = RandomBasedUniqueIdentifierGenerator
+    )
+    val requestContext = modifyRequestContext apply MockRequestContext.filterableMulti
+      .copy(
+        indexPacks = indexPacks,
+        action = Action("indices:data/read/mget"),
+        isReadOnlyRequest = true,
+        method = Method("POST"),
+        allIndicesAndAliases = Set(
+          fullLocalIndexWithAliases(fullIndexName("test1"), Set.empty),
+          fullLocalIndexWithAliases(fullIndexName("test2"), Set.empty),
+          fullLocalIndexWithAliases(fullIndexName("test3"), Set.empty),
+          fullLocalIndexWithAliases(fullIndexName("test4"), Set.empty),
+          fullLocalIndexWithAliases(fullIndexName("test5"), Set.empty)
+        )
+      )
+    val blockContext = FilterableMultiRequestBlockContext(
+      requestContext,
+      UserMetadata.from(requestContext),
+      Set.empty,
+      List.empty,
+      indexPacks,
+      None
+    )
+    rule.check(blockContext).runSyncUnsafe() shouldBe {
+      if (isMatched) {
+        Fulfilled(FilterableMultiRequestBlockContext(
+          requestContext,
+          UserMetadata.from(requestContext),
+          Set.empty,
+          List.empty,
+          allowed,
+          None
+        ))
+      } else {
+        Rejected(Some(Cause.IndexNotFound))
+      }
+    }
+  }
+
+  protected def assertMatchRuleForTemplateRequest(configured: NonEmptySet[RuntimeMultiResolvableVariable[ClusterIndexName]],
+                                                requestContext: MockTemplateRequestContext,
+                                                templateOperationAfterProcessing: TemplateOperation,
+                                                allAllowedIndices: Set[ClusterIndexName],
+                                                additionalAssertions: TemplateRequestBlockContext => Assertion = noTransformation): Assertion = {
+    val rule = createIndicesRule(configured)
+    val ruleResult = rule.check(requestContext.initialBlockContext).runSyncStep.toOption.get
+    ruleResult should matchPattern {
+      case Fulfilled(blockContext@TemplateRequestBlockContext(rc, metadata, headers, Nil, operation, _, allowedIndices))
+        if rc == requestContext
+          && metadata == requestContext.initialBlockContext.userMetadata
+          && headers.isEmpty
+          && operation == templateOperationAfterProcessing
+          && allowedIndices == allAllowedIndices
+          && additionalAssertions(blockContext) == Succeeded =>
+    }
+  }
+
+  protected def assertNotMatchRuleForTemplateRequest(configured: NonEmptySet[RuntimeMultiResolvableVariable[ClusterIndexName]],
+                                                   requestContext: MockTemplateRequestContext,
+                                                   specialCause: Option[Cause] = None): Assertion = {
+    val rule = createIndicesRule(configured)
+    val ruleResult = rule.check(requestContext.initialBlockContext).runSyncStep.toOption.get
+    ruleResult shouldBe Rejected(specialCause)
+  }
+
+  private def createIndicesRule(configuredValues: NonEmptySet[RuntimeMultiResolvableVariable[ClusterIndexName]]) = {
+    new IndicesRule(
+      settings = IndicesRule.Settings(configuredValues, mustInvolveIndices = false),
+      identifierGenerator = (_: Refined[Int, Positive]) => "0000000000"
+    )
+  }
+
+  protected def indexNameVar(value: NonEmptyString): RuntimeMultiResolvableVariable[ClusterIndexName] = {
+    variableCreator
+      .createMultiResolvableVariableFrom(value)(AlwaysRightConvertible.from(clusterIndexName))
+      .getOrElse(throw new IllegalStateException(s"Cannot create IndexName Value from $value"))
+  }
+
+  protected implicit class MockTemplateRequestContextOps(underlying: MockTemplateRequestContext) {
+    def addExistingTemplates(template: Template, otherTemplates: Template*): MockTemplateRequestContext = {
+      underlying.copy(allTemplates = underlying.allTemplates + template ++ otherTemplates.toSet)
+    }
+  }
+
+  private def noTransformation(blockContext: TemplateRequestBlockContext) = {
+    // we check here if sth else than identity was configured
+    val controlTemplates: Set[Template] = Set(
+      LegacyTemplate(TemplateName("whatever1"), UniqueNonEmptyList.of(indexPattern("*")), Set(clusterIndexName("alias"))),
+      IndexTemplate(TemplateName("whatever2"), UniqueNonEmptyList.of(indexPattern("*")), Set(clusterIndexName("alias"))),
+      ComponentTemplate(TemplateName("whatever3"), Set(clusterIndexName("alias"))),
+    )
+    blockContext.responseTemplateTransformation(controlTemplates) should be(controlTemplates)
+  }
+
+  protected def fullRemoteIndexWithAliases(clusterName: String,
+                                         fullRemoteIndexName: String,
+                                         remoteIndexAliases: String*): FullRemoteIndexWithAliases = {
+    def fullIndexNameFrom(value: String) = {
+      IndexName.Full.fromString(value) match {
+        case Some(name) => name
+        case _ => throw new IllegalArgumentException(s"Cannot create full index name from '$value'")
+      }
+    }
+
+    FullRemoteIndexWithAliases(
+      ClusterName.Full.fromString(clusterName).getOrElse(throw new IllegalArgumentException(s"Cannot create cluster name from '$clusterName'")),
+      fullIndexNameFrom(fullRemoteIndexName),
+      IndexAttribute.Opened,
+      remoteIndexAliases.toSet.map(fullIndexNameFrom)
+    )
+  }
+
+  protected def fullRemoteDataStreamWithAliases(clusterName: String,
+                                                fullRemoteDataStreamName: NonEmptyString,
+                                                remoteDataStreamAliases: NonEmptyString*): FullRemoteDataStreamWithAliases = {
+    FullRemoteDataStreamWithAliases(
+      ClusterName.Full.fromString(clusterName).getOrElse(throw new IllegalArgumentException(s"Cannot create cluster name from '$clusterName'")),
+      fullDataStreamName(fullRemoteDataStreamName),
+      aliasesNames = remoteDataStreamAliases.toSet.map(fullDataStreamName),
+      backingIndices = Set(fullIndexName(NonEmptyString.unsafeFrom(".ds-" + fullRemoteDataStreamName.value)))
+    )
+  }
+
+  private val variableCreator: RuntimeResolvableVariableCreator =
+    new RuntimeResolvableVariableCreator(TransformationCompiler.withAliases(SupportedVariablesFunctions.default, Seq.empty))
+}
